Fix incomplete install caused by incorrect exit condition.
s:install() spawns up to g:neobundle#install_max_processes processes
at a time and waits for them to finish. These processes are executed
synchronously when vimproc is not used, so the loop's exit condition
"empty(context.source__processes)" is always true at the end of each
iteration. This causes the installation to stop too early while some
bundles are yet to be installed. This can probably also happen when
vimproc is used, if all processes happen to finish at the same time.

This bug probably appeared due to 436e4d4. Prior to that commit, the
exit condition was checked before pruning finished processes from the
list. This causes an extra loop iteration (which is what the commit
intended to fix), but as a side effect the correct exit condition is
checked again and installation continues until all bundles have been
installed.