Add verbose message for FileType handling
It is useful to know that Neobundle resets and/or triggers the FileType
autocommand.  This patch adds a message in case &verbose is set, and
displays the bundles in the context.

It could be improved to only display the bundle(s) that triggered the
resetting, but it's probably not worth it and not so easy for when
"only" the `FileType` autocommand is triggered.  In that case a diff of
the output from `autocmd FileType` might be useful.