neobundle#commands#load_cache: skip non-existent vimrc files
With `vim -u foo.vim`, `$MYVIMRC` (the default for `a:vimrcs`) is not
set/empty and cannot be used to check the cache's freshness: the cache
will be considered to be fresh always currently.

But overwriting an existing cache should also be prevented in this case!
Otherwise you will create a new cache file (from some custom vimrc),
which is then possibly newer than your vimrc.